build: Rework to use nonrecursive makefile.
This largely keeps the same configure interface, that is -
./autogen.sh for the first time and then ./configure --enable-foo
--with-bar=baz && make.

There is new way of running functional tests - ./configure
--enable-functional-tests && make check

Removes the generation of aggregate script in favor of using make to
install files into ACI rootfs.

Adds dependency tracking for binaries (both go and C).

See rkt-bs-todo for a list of things that are left to be done.