build: get rkt binary linked with static libs by default
Introduce a configure option --enable-shared to be able to support
both shared library linking and static lib linking. Default is 'no',
which means a static-lib-only mode.
Doing that, rkt binary can be linked with static libraries by default,
so that rkt can be installed and distributed for distros with different
combinations of system libraries.

Fixes: https://github.com/coreos/rkt/issues/1344